requirements.txt
fpdf==1.7.2
Pillow==10.0